Split breakpoints, callstack, sources files
